Move test fixtures into `conftest.py`
