+++ examples/boutmodules/tests/test_stormdataset.py
@@ -2,7 +2,6 @@ import pytest
 
 from xarray import concat
 from xbout import open_boutdataset
-from xbout.tests.test_load import bout_xyt_example_files
 
 
 @pytest.mark.skip
